interfaces/builtin/qualcomm-ipc-router: rename from qrtr and add more details
